fix: no longer endless loop when calling nextStep in callback
Before this commit if you tried to call nextStep from within the
onNextStep callback this would always result in an endless loop.
Incrementing the current step first solves this problem.

An example case this impacts is if you wanted to skip a step based on a
condition in your onNextStep callback.

ref: https://github.com/pulsardev/vue-tour/issues/82